Move bees module TileEntity to its proper sub directory
