+++ src/java/growthcraft/bees/gui/GuiBeeBox.java
@@ -1,7 +1,7 @@
 package growthcraft.bees.gui;
 
 import growthcraft.bees.block.ContainerBeeBox;
-import growthcraft.bees.entity.TileEntityBeeBox;
+import growthcraft.bees.tileentity.TileEntityBeeBox;
 
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.client.resources.I18n;
